+++ examples/nextjs-app/config.ts
@@ -4,7 +4,7 @@ import { beamTestnet, polygonAmoy, sepolia } from 'viem/chains';
 
 export const config = createConfig(
   getDefaultConfig({
-    appName: 'OpenfortKit Next.js demo',
+    appName: 'Openfort Next.js demo',
     chains: [beamTestnet, polygonAmoy, sepolia],
     walletConnectProjectId: process.env.NEXT_PUBLIC_WALLET_CONNECT_PROJECT_ID!,
   })