+++ packages/openfort-react/src/components/Openfort/useOpenfort.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { OpenfortKitContext } from './context';
+import { Openfortcontext } from './context';
 
 
 export const useOpenfort = () => {
-  const context = React.useContext(OpenfortKitContext);
-  if (!context) throw Error('OpenfortKit Hook must be inside a Provider.');
+  const context = React.useContext(Openfortcontext);
+  if (!context) throw Error('Openfort Hook must be inside a Provider.');
   return context;
 };